LPD-69776 Rename useOverlayPositon to useOverlayPosition
